feat: authorization and user in session working

+++ pages/api/bikes/[id].ts
@@ -2,6 +2,7 @@ import type { NextApiRequest, NextApiResponse } from 'next';
 
 import dbConnect from '../../../utils/dbConnect';
 import Bike from '../../../models/Bike';
+import { getSession } from 'next-auth/client';
 
 export default async function handler(
   req: NextApiRequest,
@@ -11,6 +12,23 @@ export default async function handler(
     query: { id },
     method,
   } = req;
+
+  const session = await getSession({ req });
+
+  const checkAccess = () => {
+    console.log('Check access');
+    if (!session) {
+      // Signed in
+      res.status(401).json('401 - Unauthorized. Please log in');
+      res.end();
+      return false;
+    } else if (session.user.role !== 'manager') {
+      res.status(403).json('403 - Forbidden. Not enough rights');
+      res.end();
+      return false;
+    }
+    return true;
+  };
 
   await dbConnect();
 
@@ -29,6 +47,7 @@ export default async function handler(
 
     case 'PUT':
       try {
+        if (!checkAccess()) return;
         let bike = await Bike.findOne({ _id: id });
 
         const { model, color, location, rating, image } = JSON.parse(req.body);
@@ -51,8 +70,9 @@ export default async function handler(
       }
       break;
 
-    case 'DELETE' /* Delete a model by its ID */:
+    case 'DELETE':
       try {
+        if (!checkAccess()) return;
         const deletedBike = await Bike.deleteOne({ _id: id });
         if (!deletedBike) {
           return res.status(400).json({ success: false });
